fix: updated card control

+++ src/components/Card/Card.jsx
@@ -1,24 +1,49 @@
-import { deleteTask, updateTask } from "../../service/task";
+import { deleteAllTasks, deleteTask, updateAllTasks, updateTask } from "../../service/task";
 import { formatLocalDate } from "../../helpr/date";
 
 import IconDelete from "../../assets/icons/delete.svg";
-import IconCheck from "../../assets/icons/check.svg";
+import IconEdit from "../../assets/icons/edit.svg";
 
 import "./Card.css";
 
-export default ({ id, title, content, date, done, update }) => {
+export default ({ id, title, content, date, update, hourStart, hourEnd }) => {
+
   async function handleDeleteTask() {
-    await deleteTask(id);
-    update();
+    const confirm = window.confirm("Deseja excluir?");
+    if (confirm) {
+      await deleteTask(id);
+      await deleteAllTasks(id);
+      update();
+    }
   }
 
-  async function handleCompleteTask() {
-    await updateTask(id, { done: true });
+  async function handleEditTask() {
+    const titleEdit = window.prompt("Título:", String(title));
+    if (titleEdit == null) return;
+
+    const contentEdit = window.prompt("Descrição:", String(content));
+    if (contentEdit == null) return;
+
+    const hourStartEdit = window.prompt("Início (HH:MM):", String(hourStart));
+    if (hourStartEdit == null) return;
+
+    const hourEndEdit = window.prompt("Término (HH:MM):", String(hourEnd));
+    if (hourEndEdit == null) return;
+
+    const data = {
+      title: titleEdit,
+      content: contentEdit,
+      hourStart: hourStartEdit,
+      hourEnd: hourEndEdit
+    };
+
+    await updateTask(id, data);
+    await updateAllTasks(id, data);
     update();
   }
 
   return (
-    <div className={done ? "card done" : "card"}>
+    <div className="card">
       <small className="date">
         {formatLocalDate(date)}
       </small>
@@ -29,14 +54,12 @@ export default ({ id, title, content, date, done, update }) => {
       </div>
 
       <div className="control">
-        {!done && (
-          <button onClick={handleCompleteTask} className="button" data-title="Concluir">
-            <img src={IconCheck} className="icon" />
-          </button>
-        )}
-
         <button onClick={handleDeleteTask} className="button" data-title="Remover">
           <img src={IconDelete} className="icon" />
+        </button>
+
+        <button onClick={handleEditTask} className="button" data-title="Editar">
+          <img src={IconEdit} className="icon" />
         </button>
       </div>
     </div>
